inference_service/requirements.txt
pandas
scikit-learn
joblib
paho-mqtt
numpy